requirements.txt
streamlit==1.40.0
torch==2.5.1
transformers==4.46.2
sentence-transformers==3.3.1
numpy==1.26.4
faiss-cpu==1.9.0
rank-bm25==0.2.2